+++ src/VsixTesting/Vs/VisualStudioUtil.cs
@@ -8,80 +8,15 @@ namespace Vs
     using System.IO;
     using System.Linq;
     using System.Reflection;
-    using System.Runtime.InteropServices;
     using System.Threading.Tasks;
     using Common;
-    using EnvDTE80;
     using Microsoft.VisualStudio.Setup.Configuration;
     using Microsoft.Win32;
     using VsixTesting.Utilities;
-    using DTE = EnvDTE.DTE;
 
-    internal static class VisualStudioUtil
+    internal static partial class VisualStudioUtil
     {
         public const string ProcessName = "devenv";
-
-        public static DTE GetDTE(Process process)
-        {
-            string namePattern = $@"!VisualStudio\.DTE\.(\d+\.\d+):{process.Id.ToString()}";
-            return (DTE)RunningObjectTable.GetRunningObjects(namePattern).FirstOrDefault();
-        }
-
-        public static async Task<DTE> GetDTE(Process process, TimeSpan timeout)
-        {
-            var msTimeout = timeout.TotalMilliseconds;
-            while (true)
-            {
-                var dte = GetDTE(process);
-                if (dte != null)
-                    return dte;
-                await Task.Delay(250);
-                if ((msTimeout -= 250) <= 0)
-                    throw new TimeoutException($"Failed getting DTE from {process.ProcessName} after waiting {timeout.TotalSeconds} seconds");
-            }
-        }
-
-        public static IEnumerable<DTE> GetRunningDTEs()
-        {
-            foreach (var process in Process.GetProcessesByName(ProcessName))
-            {
-                if (GetDTE(process) is DTE dte)
-                    yield return dte;
-            }
-        }
-
-        public static DTE GetDteFromDebuggedProcess(Process process)
-        {
-            foreach (var dte in GetRunningDTEs())
-            {
-                try
-                {
-                    if (dte.Debugger.CurrentMode != EnvDTE.dbgDebugMode.dbgDesignMode)
-                    {
-                        foreach (Process2 debuggedProcess in dte.Debugger.DebuggedProcesses)
-                        {
-                            if (debuggedProcess.ProcessID == process.Id)
-                                return dte;
-                        }
-                    }
-                }
-                catch (COMException)
-                {
-                    continue;
-                }
-            }
-
-            return null;
-        }
-
-        public static void AttachDebugger(DTE dte, Process targetProcess, string engine = "Managed")
-        {
-            dte?.Debugger
-                .LocalProcesses
-                .Cast<Process2>()
-                .FirstOrDefault(p => p.ProcessID == targetProcess.Id)
-                ?.Attach2(engine);
-        }
 
         public static IEnumerable<VsInstallation> FindInstallations()
         {
